Changed suiteId to integer value
Changed the testObjectSuiteId value in the annotation with an integer value. Before, this was an undefined variable. Since our custom runner can read suiteId via environment variables, this allows us to skip editing the code during demos: we will just be able to export the environment variables (as per usual procedure) and run mvn test.